refactor(frontend): remove dead WorkspaceNav sidebar remnants (#133)
Move WorkspaceNavItem/DashboardWorkspaceKey/WorkspaceNavAction types into WorkspaceTopNav (the live Command nav), retarget App.tsx/phytosyncSections.ts/WorkspaceTopNav.test.tsx imports, and delete the unused WorkspaceNav sidebar, its test, the phyto/sidebar re-export, and ui/sidebar (blue-gradient glass shells off the Command palette). Closes parity ledger rows 12, 13, 67.
