--- frontend/src/components/phyto/sidebar/Sidebar.tsx
@@ -1,1 +1,0 @@
-export { default } from '../../shell/WorkspaceNav';
